requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.0
python-dotenv==1.0.1
google-generativeai==0.7.2
qdrant-client==1.9.1